Add files via upload
This project dynamically fetches and displays information about the Harry Potter film series using JavaScript. It allows users to select a movie from a dropdown menu and view details like the director, writer, producer, music, poster, and trailer. The data is fetched from a local JSON file and displayed in an interactive interface.